Home screen refactor
