+++ constants/placeholderConstants.tsx
@@ -1,0 +1,2 @@
+export const RIDE_MODE = 'Get a Ride';
+export const POOL_MODE = 'Pool My Ride';
